+++ core/frontend/services/themes/ThemeStorage.js
